created login and signup apis with UI and jwt authentication

+++ backend/db/migrations/20241018131635_blog_table.js
@@ -3,13 +3,20 @@
  * @returns { Promise<void> }
  */
 exports.up = function(knex) {
-  return knex.schema.createTable('blogs', (table) => {
+  return knex.schema.createTableIfNotExists('blogs', (table) => {
     table.increments();
     table.string('title').notNullable();
     table.string('preview').notNullable();
     table.string('post').notNullable();
     table.timestamp('created_at').defaultTo(knex.fn.now());
     table.timestamp('updated_at').defaultTo(knex.fn.now());
+  })
+  .createTable('users', (table) => {
+    table.increments();
+    table.string('name').notNullable();
+    table.string('email').unique().notNullable();
+    table.string('password').notNullable();
+    table.string('role').notNullable().defaultTo('public');
   })
 };
 
@@ -18,5 +25,15 @@ exports.up = function(knex) {
  * @returns { Promise<void> }
  */
 exports.down = function(knex) {
-  return knex.schema.dropTable('blogs');
+  return knex.schema.hasTable('blogs').then(function(exists) {
+    if (exists) {
+      return knex.schema.dropTable('blogs');
+    }
+  }).then(() => {
+    return knex.schema.hasTable('users').then(function(exists) {
+      if (exists) {
+        return knex.schema.dropTable('users');
+      }
+    });
+  });
 };
